+++ Hues/main.m
@@ -1,0 +1,14 @@
+//
+//  main.m
+//  Hues
+//
+//  Created by Zach Waugh on 5/30/11.
+//  Copyright 2011 Giant Comet. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+int main(int argc, char *argv[])
+{
+  return NSApplicationMain(argc, (const char **)argv);
+}